# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation des API disponibles pour gérer les articles de notre application",
        "version": "0.1.1",
        "title": "Demo Blog",
        "contact": {
            "email": "tricard.sylvain@gmail.com"
        }
    },
    "host": "localhost:8000",
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/articles/{articleId}": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "Récupère les infos d'un article",
                "description": "Retourne les données d'un seul article",
                "operationId": "getArticleById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "Id de l'article",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Article"
                        }
                    },
                    "404": {
                        "description": "Ressource inexistante",
                        "schema": {
                            "$ref": "#/definitions/404_ResourceNotFound"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "articles"
                ],
                "summary": "Modifie un article existant",
                "description": "",
                "operationId": "updateArticle",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "Id de l'article",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "Titre de l'article",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "contents",
                        "in": "formData",
                        "description": "Contenu de l'article",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "status_id",
                        "in": "formData",
                        "description": "Id du statut de l'article",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "publication_date",
                        "in": "formData",
                        "description": "Date de publication",
                        "required": false,
                        "type": "string",
                        "format": "datetime"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Article"
                        }
                    },
                    "400": {
                        "description": "Paramètres invalides",
                        "schema": {
                            "$ref": "#/definitions/400_ValidationError"
                        }
                    },
                    "404": {
                        "description": "Ressource inexistante",
                        "schema": {
                            "$ref": "#/definitions/404_ResourceNotFound"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "articles"
                ],
                "summary": "Passe l'article en statut \"SUPPRIME\"",
                "description": "",
                "operationId": "deleteArticle",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "Id de l'article",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Paramètres invalides",
                        "schema": {
                            "$ref": "#/definitions/400_ValidationError"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "api-key",
            "in": "header"
        }
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Sylvain"
                },
                "email": {
                    "type": "string",
                    "example": "tricard.sylvain@gmail.com"
                }
            },
            "xml": {
                "name": "User"
            }
        },
        "Article": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Comment documenter une API ?"
                },
                "contents": {
                    "type": "string",
                    "example": "lorem ipsum"
                },
                "publicated_at": {
                    "type": "string",
                    "format": "datetime",
                    "example": "2021-09-14 18:00"
                },
                "user": {
                    "$ref": "#/definitions/User"
                },
                "article_status": {
                    "$ref": "#/definitions/ArticleStatus"
                }
            },
            "xml": {
                "name": "Article"
            }
        },
        "ArticleStatus": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "code": {
                    "type": "string",
                    "example": "BROUILLON"
                },
                "label": {
                    "type": "string",
                    "example": "brouillon"
                }
            },
            "xml": {
                "name": "ArticleStatus"
            }
        },
        "400_ValidationError": {
            "description": "Your request parameters didn't validate.",
            "required": [
                "status",
                "type",
                "title",
                "invalid-params"
            ],
            "properties": {
                "type": {
                    "type": "string",
                    "example": "problems/validation-error"
                },
                "title": {
                    "type": "string",
                    "example": "Your request parameters didn't validate."
                },
                "status": {
                    "type": "integer",
                    "format": "int32",
                    "example": "400"
                },
                "invalid-params": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/InvalidParams"
                    }
                }
            }
        },
        "InvalidParams": {
            "description": "Invalid params list.",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "email"
                },
                "reason": {
                    "type": "string",
                    "example": "must be a valid email address"
                }
            }
        },
        "404_ResourceNotFound": {
            "description": "The resource you looking for can't be found.",
            "required": [
                "type",
                "title",
                "status",
                "detail"
            ],
            "properties": {
                "type": {
                    "type": "string",
                    "example": "/problems/resource-not-found"
                },
                "title": {
                    "type": "string",
                    "example": "The resource you looking for can't be found."
                },
                "status": {
                    "type": "integer",
                    "format": "int32",
                    "example": 404
                },
                "detail": {
                    "type": "string",
                    "example": "[Model] [id] appear to be not existing "
                }
            }
        },
...
    },
...
}